requirements.txt
vosk
openai-whisper>=0.3.45
